app/schemas/skill.py
from pydantic import BaseModel, ConfigDict, Field


class SkillBase(BaseModel):
    name: str = Field(min_length=1, max_length=255)


class SkillCreate(SkillBase):
    pass


class SkillUpdate(BaseModel):
    name: str | None = Field(default=None, min_length=1, max_length=255)


class SkillRead(SkillBase):
    model_config = ConfigDict(from_attributes=True)

    id: int
